Users,Subjects,SubjectSources,SubjectTypes
http://localhost/archon/?p=core/subjects&apilogin=admin&apipassword=admin&substart=1
This URL Returns subjects in batches of 100 . Substart decides the
beginning marker. If the substart is 1 then the group returned is  1-100
(0  is handles as 1 as well) , the next substart should be incremented
by 100.101 -201 would be the next range.

http://localhost/archon/?p=core/users&apilogin=admin&apipassword=admin
This URL Returns a list of users the permissions array for each object
contains the relationship between users the permissions and their
repositories .

http://localhost/archon/?p=core/subjectsources&apilogin=admin&apipassword=admin
This URL Returns a list of Subject sources

http://localhost/archon/?p=core/subjecttypes&apilogin=admin&apipassword=admin
This URL Returns a list subjecttypes.