enums,creators,locations
http://localhost/archon/?p=core/creators&apilogin=admin&apipassword=admin&enum_type=creators&batch_start=1

This url returns a list of creators in batches of 100.

http://localhost/archon/?p=core/enums&apilogin=admin&apipassword=admin&enum_type=creatorsources
this Url returns a list of creator name sources

http://localhost/archon/?p=core/enums&apilogin=admin&apipassword=admin&enum_type=extentunits
This url returns a list of units.

http://localhost/archon/?p=core/enums&apilogin=admin&apipassword=admin&enum_type=filetypes
This url returns a list of file types.

http://localhost/archon/?p=core/enums&apilogin=admin&apipassword=admin&enum_type=levelcontainers
This url returns a list of containers.

http://localhost/archon/?p=core/enums&apilogin=admin&apipassword=admin&enum_type=materialtypes
This url returns a list of material types.

http://localhost/archon/?p=core/enums&apilogin=admin&apipassword=admin&enum_type=descriptiverules
This url shows descriptive rules.

http://localhost/archon/?p=core/locations&apilogin=admin&apipassword=admin
This URL shows Locations.